Define JUnit 5 dependency and update strategy
So far, the project depended on JUnit 5's SNAPSHOT versions, which
allowed for care-free development (particularly before JUnit 5's GA)
but releases obviously have to depend on released JUnit 5 versions.
This made it necessary to define a general strategy for handling the
dependency on JUnit 5.

Developing against SNAPSHOT versions and only picking a released
version when Pioneer gets released has some appeal. It leads to an
unreleasable `master`, though, because code may depend on unreleased
JUnit 5 APIs. Then, how do we ship bug fixes or smaller improvements?
Cutting release branches or merging features based on unreleased APIs
into a "shadow master" adds a lot of complexity.

In the end, the benefits of depending on SNAPSHOTs do not outweigh
the cost. It is easier to limit ourselves to developing against
released versions only. If push comes to shove, a "shadow master" can
still be created, and if more immediate feedback regarding
JUnit-5-internal changes is desired, it would be easy to add another
build job that builds against SNAPSHOTs.

Issue: #64
PR: #90