Document Pioneer versioning by contributors/Shipkit
The documentation already stated that Pioneer uses Shipkit, but
didn't explain how it created versions or how Pioneer wants
to handle versioning.

In an effort not to repeat Shipkit's documentation, its process isn't
explained in detail. The main addition is to define that contributors
need to discuss a branch's impact on Pioneer in the PR, so they can
decide there whether to bump the minor version.

Closes: #113
PR: #150