Add extension to change system environment variables
The `@ClearEnvironmentVariable` and `@SetEnvironmentVariable`
annotations can be used to clear, respectively, set the values
of environment variables for a test execution. Because the JVM
stores environment variables as an immutable copy of the
operating system's variables, we have to sue reflection to
change them. Appropriate warnings are added to the documentation
and written to the log.

PR: #174
Closes: #167