Ensure OS independence of EnvironmentVariableUtils (#287 / #288)
The method EnvironmentVariableUtils.modifyEnvironmentVariables calls
a method which works on Windows first and then a method which works
on Linux and OS X. The former accesses the field
java.lang.ProcessEnvironment.theEnvironment by reflection, assumes it
to be a Map<String, String>, and modifies it. On at least some
platforms, java.lang.ProcessEnvironment.theEnvironment exists but is
a Map<Variable, Value>, so the map is corrupted after being modified.
Subsequent calls to fetch data out of System.getenv() result in a
ClassCastException.

This change ensures that no changes take place in the Windows variant
until all reflection operations have completed. If any reflection
operation fails, the alternate Linux/Mac OS approach is used without
having modified the environment.

closes: #287
PR: #288