Include `@xp:version` in project-page.adoc
The excerpt of the "About Pioneer" page should include the project
version as described in junit-pioneer/junit-pioneer.github.io#9.

This commit also updates the description to the current version
that evolved over time in the website's `home.md`. To prevent future
diversion between the pages, the website will pull in this excerpt
when building the landing page.