Configure Git to normalize line endings to LF (#341 / #347)
As observed in #341, without further configuration, Git uses the OS
default line ending (LF on Unix, CRLF on Windows) when writing files
to the working directory, e.g. during checkout. If other tools aren't
aware of that and edit line endings, the resulting diff is noisy.

Specifically this happened in #341, where Pioneer was checked out on
Windows (~> CRLF) and Spotless changed all line endings (to LF).

This change creates a `.gitattributes` file with the following
settings:

# all end-of-lines are normalized to LF when written to the repository
# https://git-scm.com/docs/gitattributes#_text
* text=auto

# force all text files on the working dir to have LF line endings
# https://git-scm.com/docs/gitattributes#_eol
* text eol=lf

Closes: #341
PR: #347