Remove @TempDir extension (#277 / #327)
A very similar extension (with the same name) exists in Jupiter. It
doesn't support custom file systems, but other than that is more
feature-rich than this variant. This creates an unpleasant situation
where there are two extensions that solve the same problem with the
same approach, but slightly different feature sets.

After evaluating the situation, we decided to remove our extension
before the 1.0 release with the goal to look for a better solution
to the problem in the future.

Closes: #277
Related: #348
PR: #327