Add `minSuccess` to `@RetryingTest` (#408 / #430)
Allows the user to specify an attribute `minSuccess` on
`@RetryingTest` that requires that the retried test succeeds at least
N times. For better readability, `minSuccess` is mirrored by a new
attribute `maxAttempts`. For better usability, the existing attribute
`value` becomes an alias for `maxAttempts` - it is mandatory to set
exactly one of the two.

Closes: #408
PR: #430