Update to JUnit Jupiter 5.7 (#431 / #445)
We usually want to depend on the lowest minor version (newest patch
of that) that supports our feature set. Generally speaking, that's
still the case, but the upcoming hack.commit.push event required an
exception:

Updating from Jupiter 5.6 to 5.7 requires some changes to our test
code, which means it's a bit disruptive. In case some contributors
need to update during the event, they would also need to make these
changes in our test code, which would be an unnecessary burden. To
preempt that, I decided to update to 5.7 now.

Relates to: #431
PR: #445