Refactor `PioneerAnnotationUtils` and tests (#457)
The hunt after build timeout causes (see #324) implicated
`PioneerAnnotationUtilsTests`. Removing the Escheresque indirection
and replacing the tests with a simpler approach unearthed some bugs
and missing cases in `PioneerAnnotationUtils`. This change fixes them
and seemingly also fixed the timeouts. We're so positive about that
that we even turn the Windows builds back on.

Relates to: #324
PR: #457