Define version in GitHub release action (#442 / #466)
Shipkit auto-versioning-plugin can detect the version to be released
on its own, but it increases the patch versions by number of commits
since recent release (hence 1.3.0 ~> 1.3.8), which is not what we
want. A feature request for incremental increases was declined.[1]

## Options

We hence need to define the version externally. The two main
options are:

1. pushing a tag and detecting that from Gradle
2. defining it in the manual GitHub Action[2] and passing it to Gradle

We didn't discuss preference as 1. does currently not work with
Shipkit because the changelog plugin fails to fill in the GitHub
release.[3]

## GitHub Action

We went with 2., which requires a slight change to the build and a
larger change to the documentation in `CONTRIBUTING.md`.

## `version.properties` file

Contrary to what the documentation seems to suggest, the file
`version.properties` and the `version` property are always needed,
even if the property is specified on the command line. We hence keep
the file. In the hope that we don't have to keep updating it, we set
a bogus version, though. 🤞🏾

[1] https://github.com/shipkit/shipkit-auto-version/issues/72
[2] https://github.blog/changelog/2020-07-06-github-actions-manual-triggers-with-workflow_dispatch/
[3] https://github.com/shipkit/shipkit-changelog/issues/97

Closes: #442
PR: #466