Create optional dependency policy (#502 / #593)
One of Pioneer's cornerstones is not to add to user's dependency hell
and so we don't want to require any run-time dependencies beyond
JUnit (and its dependencies). There are features that we can't
(reasonably) implement without additional dependencies, though:

* integration with other frameworks (e.g. Playwright, see #426)
* ease of use for complex functionality (e.g. JSON parsing, see #101)

There are various options to take on such dependencies and different
lenses through which to view them - for details see #502. In the end
we decided to add optional dependencies to the main project because:

* while the impact on user's dependency hell isn't zero, it should
  be reasonably close to that
* it has the least complexity for us when it comes to setting up and
  maintaining different subprojects
* the features are easy to discover by users

Closes: #502
PR: #593