Configure build and docs to compile and test code snippets (#335 / #334)
Code snippets in the documentation files in `docs` are just text and
thus run the risk of getting outdated when APIs evolve or
implementations change their behavior. To address that, we now use
Asciidoc's capability to include tagged regions from external
files[1] that are compiled and tested by the build. These files'
primary purpose is to demonstrate features (not to test them) and so
they are placed in a new source set `src/demojava`.

[1]: https://docs.asciidoctor.org/asciidoc/latest/directives/include-tagged-regions/

Related to junit-pioneer/junit-pioneer.github.io#33
Closes: #335
PR: #334